Move input visitor helper to Form component utilities
