+++ client/src/mvc/form/form-data.js
@@ -1,7 +1,7 @@
 import _ from "underscore";
 import $ from "jquery";
 import Backbone from "backbone";
-import { getGalaxyInstance } from "app";
+import { visitInputs } from "components/Form/utilities";
 
 /* This class maps the form dom to an api compatible javascript dictionary. */
 export var Manager = Backbone.Model.extend({
@@ -200,50 +200,6 @@ export var matchCase = (input, value) => {
         }
     }
     return -1;
-};
-
-/** Visits tool inputs
- * @param{dict}   inputs    - Nested dictionary of input elements
- * @param{dict}   callback  - Called with the mapped dictionary object and corresponding model node
- */
-export var visitInputs = (inputs, callback, prefix, context) => {
-    context = $.extend({}, context);
-    _.each(inputs, (input) => {
-        if (input && input.type && input.name) {
-            context[input.name] = input;
-        }
-    });
-    for (var key in inputs) {
-        const Galaxy = getGalaxyInstance();
-        var node = inputs[key];
-        node.name = node.name || key;
-        var name = prefix ? `${prefix}|${node.name}` : node.name;
-        switch (node.type) {
-            case "repeat":
-                _.each(node.cache, (cache, j) => {
-                    visitInputs(cache, callback, `${name}_${j}`, context);
-                });
-                break;
-            case "conditional":
-                if (node.test_param) {
-                    callback(node.test_param, `${name}|${node.test_param.name}`, context);
-                    var selectedCase = matchCase(node, node.test_param.value);
-                    if (selectedCase != -1) {
-                        visitInputs(node.cases[selectedCase].inputs, callback, name, context);
-                    } else {
-                        Galaxy.emit.debug(`form-data::visitInputs() - Invalid case for ${name}.`);
-                    }
-                } else {
-                    Galaxy.emit.debug(`form-data::visitInputs() - Conditional test parameter missing for ${name}.`);
-                }
-                break;
-            case "section":
-                visitInputs(node.inputs, callback, name, context);
-                break;
-            default:
-                callback(node, name, context);
-        }
-    }
 };
 
 export default {
